refactor(Utils): move price tag handling in the proof utils
