+++ src/views/PriceValidationAssistant.vue
@@ -49,8 +49,8 @@ import { defineAsyncComponent } from 'vue'
 import constants from '../constants'
 import api from '../services/api'
 import { useAppStore } from '../store'
-import barcode_utils from '../utils/barcode.js'
 import date_utils from '../utils/date.js'
+import proof_utils from '../utils/proof.js'
 import utils from '../utils.js'
 
 export default {
@@ -172,61 +172,14 @@ export default {
           for (let i = 0; i < data.items.length; i++) {
             // only validate price tags with predictions
             if (data.items[i]['predictions'].length > 0) {
-              const priceTagPrediction = data.items[i]['predictions'][0]
-              const label = priceTagPrediction['data']
-              const barcodeString = label.barcode ? barcode_utils.cleanBarcode(label.barcode.toString()) : ''
-
-              // The first schema was not versioned, so if the field is missing,
-              // we assume it's schema version 1.0
-              const schemaVersion = priceTagPrediction.schema_version || '1.0'
-
-              // fields that are common to all schema versions are initialized here
-              let productPriceForm = {
-                id: data.items[i].id,
-                origins_tags: ![null, '', 'unknown', 'other'].includes(label.origin) ? [label.origin] : [],
-                currency: data.items[i]['proof'].currency || this.appStore.getUserLastCurrencyUsed,
-                proof: data.items[i]['proof'],
-                proofImage: data.items[i]['proof'].file_path,
-                croppedImage: null,
-                product_code: barcodeString,
-                detected_product_code: barcodeString,
-                product_name: label.product_name,
-                bounding_box: data.items[i].bounding_box,
-                loading: false
-              }
-
-              if (schemaVersion === '1.0') {
-                // For schema version 1.0
-                const priceType = barcodeString.length >= 8 ? constants.PRICE_TYPE_PRODUCT : constants.PRICE_TYPE_CATEGORY
-                productPriceForm.type = priceType
-                productPriceForm.category_tag = (priceType === constants.PRICE_TYPE_CATEGORY && ![null, '', 'unknown', 'other'].includes(label.product)) ? label.product : null
-                productPriceForm.labels_tags = (priceType === constants.PRICE_TYPE_CATEGORY && label.organic) ? [constants.PRODUCT_CATEGORY_LABEL_ORGANIC] : []
-                productPriceForm.price = label.price.toString()
-                productPriceForm.price_per = label.unit
-                // price_is_discounted is not supported in schema version 1.0
-                productPriceForm.price_is_discounted = false
-              } else {
-                // version 2.0 and above
-                const priceType = label.type
-                // The selected price is a price constructed by all price information available in the label,
-                // including the discount price if available.
-                const selectedPrice = label.selected_price || {}
-                productPriceForm.type = priceType
-                // we only populate category_tag and labels_tags if the price type is category
-                productPriceForm.category_tag = (priceType === constants.PRICE_TYPE_CATEGORY && ![null, '', 'unknown', 'other'].includes(label.category)) ? label.category : null
-                productPriceForm.labels_tags = (priceType === constants.PRICE_TYPE_CATEGORY && label.organic) ? [constants.PRODUCT_CATEGORY_LABEL_ORGANIC] : []
-                productPriceForm.price = selectedPrice.price ? selectedPrice.price.toString() : ""
-                productPriceForm.price_per = selectedPrice.price_per || null
-                // in schema 2.0, we detect discount information (price, type)
-                productPriceForm.price_is_discounted = selectedPrice ? selectedPrice.price_is_discounted : false
-                productPriceForm.price_without_discount = selectedPrice.price_without_discount ? selectedPrice.price_without_discount.toString() : ""
-                productPriceForm.discount_type = selectedPrice.discount_type || ""
-              }
-
-              this.productPriceForms.push(productPriceForm)
+              this.handlePriceTag(data.items[i])
             }
           }
         })
+    },
+    handlePriceTag(priceTag) {
+      let productPriceForm = proof_utils.handlePriceTag(priceTag)
+      this.productPriceForms.push(productPriceForm)
     },
     updatePriceTag(priceTagId, status, priceId) {
       return api
